requirements.txt
torch==2.1.0
torchvision==0.16.1
numpy==1.25.2
tqdm==4.66.2
scipy==1.11.4
scikit-image==0.20.0
scikit-learn==1.3.2
seaborn==0.13.0
matplotlib==3.8.1
tensorboard==2.15.2
pydicom==2.4.4
pyxu
pillow
imageio
pandas
jupyter
jupyterlab
ipywidgets